Update choose your vampire adventure.py
I added in a while loop/break stating that if the player has not received the 'victory' answer, then the program will loop back to the beginning. I also defined answer as the string being given in order to allow the victory line to register for the loop to continue/end. 3/15/2021